Add files via upload
Multi-output regression are regression problems that involve predicting two or more numerical values given an input features.

In multi-output regression, typically the outputs are dependent upon the input and upon each other. This means that often the outputs are not independent of each other and may require a model that predicts both outputs together or each output contingent upon the other outputs.